implemented and tested minimal mode and all flow of app
